[Report] Removes all possible fake re-rendering in Report and loads initially that which is visible to user
